Modularize the player container a bit more
