+++ client/src/containers/players/PlayerContainer.js
@@ -1,0 +1,43 @@
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
+
+import Spotify from './Spotify';
+
+class PlayerContainer extends Component {
+
+  play(){}
+  pause(){}
+  next(){}
+  previous(){}
+  volumeUp(){}
+  volumeDown(){}
+  mute(){}
+  unmute(){}
+
+  setPlayerInstance(){} //instantiate playerInstance for given source
+  preparePlayerInstance(){} //authorize and token setting
+  componentDidMount(){
+
+  }
+
+  render(){
+    switch(this.props.playerType) {
+      case 'spotify':
+        return <Spotify />;
+        break;
+      default:
+        return <Spotify />;
+    };
+  }
+}
+
+const mapStateToProps = (state) => ({
+  playerType: state.player.playerType,
+});
+
+PlayerContainer.propTypes = {
+  playerType: PropTypes.string,
+};
+
+export default connect(mapStateToProps)(PlayerContainer);
